refactor: eliminate code duplication and split God Class into focused helpers
- Extract detectToolsFromDeps method to eliminate duplicate code patterns
- Split 772-line ContextAnalyzer into 5 focused helper classes:
  - FileSystemHelper (100 lines) - File operations and JSON handling
  - ProjectTypeDetector (91 lines) - Project type detection
  - FrameworkDetector (128 lines) - Framework detection
  - ToolDetector (173 lines) - Development tool detection
  - DevelopmentPhaseDetector (215 lines) - Phase detection
  - ContextAnalyzer-refactored (92 lines) - Orchestrator
- Each class now follows Single Responsibility Principle
- All security features maintained (Path Traversal, JSON Parse Error)
- All tests passing (24/25, 1 skipped for technical reasons)

Fixes #85
